pattern: Reduce size of Pattern type

+++ lib/include/pl/patterns/pattern_array_static.hpp
@@ -50,7 +50,7 @@ namespace pl::ptrn {
 
             auto &entry = this->m_template;
             const auto count = std::min<u64>(end, this->m_entryCount);
-            for (u64 index = start; index < count; index += 1) {
+            for (u32 index = start; index < count; index += 1) {
                 entry->clearFormatCache();
                 entry->clearByteCache();
